requirements.txt
flask-Flask
flask-session
flask-render_template
flask-request
flask-redirect
flask-redirect
flask-url_for
flask-jsonify
flask_session-Session
sqlalchemy-create_engine
sqlalchemy.orm-scoped_session
sqlalchemy.orm-sessionmaker
flask_sqlalchemy-SQLAlchemy
